Bootstrap
Using bootstrap is making it a lot easier to do things like ading a carousel. Plus as an added bonus I get animations and responsive design for free. Need to fix the nav bar again. I need to have the Logo in the top right which can go to the instagram I guess. Also want the hamburger menu to have less of a trigger unless truely needed.